requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.2.1
django-import-export==3.0.2
django-jazzmin==2.6.0
et-xmlfile==1.1.0
filelock==3.9.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
platformdirs==2.6.2
pyperclip==1.8.2
pytz==2019.3
PyYAML==6.0
sqlparse==0.4.4
tablib==3.3.0
tzdata==2022.5
virtualenv==20.17.1
xlrd==2.0.1
xlwt==1.3.0
